Report error when there are empty display names in the CSV file
