+++ VERSION
@@ -1,1 +1,1 @@
-2.3.2
+2.4.0
