+++ exabel_data_sdk/services/csv_reader.py
@@ -3,13 +3,12 @@ from typing import Iterable, Mapping, Optional, Union
 import pandas as pd
 
 
-class CsvLoader:
-    """
-    Base class for CSV loaders which read a CSV file and upload the data to the Exabel Data API.
-    """
+class CsvReader:
+    """Reader of CSV files."""
 
+    @staticmethod
     def read_csv(
-        self, filename: str, separator: str, string_columns: Iterable[Union[str, int]]
+        filename: str, separator: str, string_columns: Iterable[Union[str, int]]
     ) -> pd.DataFrame:
         """
         Read the given file and return the content as a pandas DataFrame.
@@ -22,4 +21,4 @@ class CsvLoader:
         dtype: Optional[Mapping[Union[str, int], type]] = None
         if string_columns:
             dtype = {column: str for column in string_columns}
-        return pd.read_csv(filename, header=0, sep=separator, dtype=dtype)
+        return pd.read_csv(filename, header=0, sep=separator, dtype=dtype, keep_default_na=False)
